moved numeric types compat code to statics module
